feat: Implements Scene class
The goal of the Scene class is to handle scene management for the Game class. At this point its main duties are to handle running the update loop for scene nodes and to keep track of RenderableNodes.
